% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting_gapminder.R
\docType{data}
\name{gapminder}
\alias{gapminder}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1704 rows and 6 columns.
}
\usage{
gapminder
}
\description{
This is data to be included in my package
}
\keyword{datasets}
