% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting_mtcars.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{This is data to be included in my package}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
mtcars
}
\description{
This is data to be included in my package
}
\keyword{datasets}
