% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting_penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{This is data to be included in my package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 344 rows and 8 columns.
}
\usage{
penguins
}
\description{
This is data to be included in my package
}
\keyword{datasets}
