% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardizeAndModel.R
\name{standardize_and_model}
\alias{standardize_and_model}
\title{Standardize and model numeric vectors}
\usage{
standardize_and_model(criterion, predictor, remove_na = TRUE, ...)
}
\arguments{
\item{criterion}{the variable to be predicted; a numeric vector (note: I named this argument 'criterion' because it will be the criterion in the linear model that this function creates)}

\item{predictor}{the variable doing the predicting; a numeric vector (note: I named this argument 'predictor' because it will be the predictor in the linear model that this function creates)}

\item{remove_na}{this is a logical that evaluates to TRUE or FALSE which will indicate if you want NA values to be removed from the analysis (I used the name remove_na because when set to TRUE, NA values will be removed so the name of the argument is self-explanatory)}

\item{...}{this enables additional arguments to be passed to other functions within the function (I used the name ... because that is standard practice for the ellipsis argument)}
}
\value{
this function outputs the summary of a linear model consisting of standardized criterion and predictor variables. Note that the criterion variable will appear as 'a' and the predictor variable will appear as 'b' in the summary output.
}
\description{
Standardizes two numeric vectors and inputs the two vectors into a linear model. Particularly useful for obtaining the standardized regression coefficient of the predictor.
}
\examples{
standardize_and_model(criterion = gapminder$lifeExp, predictor = gapminder$gdpPercap) #using our new function, setting the criterion to lifeExp and predictor to gdpPercap to see if gdpPercap predicts lifeExp and obtain standardized regression coefficient etc
standardize_and_model(criterion = penguins$body_mass_g, predictor = penguins$flipper_length_mm) #using our new function by setting criterion to body_mass_g and predictor to flipper_length_mm to see if flipper length predicts body mass and obtain standardized regression coefficient etc
standardize_and_model(criterion = mtcars$mpg, predictor = mtcars$cyl) #using our new function, setting criterion to mpg and predictor to cyl to see if number of cylinders predicts miles per gallon and obtain standardized regression coefficient etc
}
