% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documenting_x.R
\docType{data}
\name{x}
\alias{x}
\title{This is data to be included in my package}
\format{
An object of class \code{summary.lm} of length 11.
}
\usage{
x
}
\description{
This is data to be included in my package
}
\keyword{datasets}
